debugging throughout if returned data pull is blank
added in if(null) throughout to avoid breaking if the returned data pull is blank (i.e. if the api call returns nothing - if there are no rows selected, it seems to return nothing, instead of a table with no rows. This was breaking various functions that used the query output.